+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All important changes to vrfy will be documented here.
 - Cli option: vrfy /path/of/master /path/of/backup no longer supported.
 - Renamed "clone" to "backup" for improved clarity.
 - Changed direction verification option -c to -b.
+- [+] and [-] cli indicators are now based on status of backup directory
 
 ### Fixed
 - Fixed bug where directories may be dropped in directory verification.